Configure journal rate-limiting to 0 when no user config
The yang data model has defaults of 10000/30s, but is
is within presence container, so might not be present.
If the data model node is not present then explicitly
mean's there should be no rate limiting.

However if journal is not configured with 0 to explicitly
disable rate limiting it implicitly defaults to rate limiting
with internal defaults values. (which just happen to match
the data model defaults)

Fixes: VRVDR-48477