Replace modelled default static journal max size
The default heuristic behaviour of journald when it not explicitly given
configuration is better for devices that are provisioned with smaller
disks. So make this behaviour explicitly available in the model.

This is technically an non backward compatible yang change, but is safe,
as journal storage is rotated out so will shrink to a smaller implicit
size in cases the devices have disks smaller than 4G, and avoids other
services like syslog, audit and journal itself running out storage to
write to.

VRVDR-54673